moves schemas into root dir of module
